requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
docutils==0.18
et-xmlfile==1.1.0
future==0.18.2
h11==0.12.0
idna==3.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.2
numpy==1.21.3
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.4
pefile==2021.9.3
Pillow==8.4.0
pycparser==2.20
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyOpenSSL==21.0.0
pyotp==2.6.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pywin32==302
pywin32-ctypes==0.2.0
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
XlsxWriter==3.0.1
